requirements.txt
fastapi
uvicorn
openai
chromadb
python-dotenv
pdfplumber
